refactor: generalize RabbitSource to multi-queue, delete RabbitBusSource
RabbitBusSource duplicated RabbitSource's basic_get + settle shape, adding
only multi-queue polling and routing-key prefix stripping. Fold those into
RabbitSource (with the sticky polling from the previous perf change) and
delete RabbitBusSource. The single-queue constructor keeps its public
signature; naming now derives from the delivery's routing key, which under
the default exchange equals the queue name, so standalone behavior is
unchanged.
